refactor: move augmentSessionsWithGarmin to lib/utils/junction-poc.ts
Isolates all Junction/Garmin PoC logic in one place instead of mixing
it into the general week-view utilities.
