+++ app/routes/athlete/week.$weekId.tsx
@@ -21,7 +21,8 @@ import { useAuth } from '~/lib/context/AuthContext';
 import { queryKeys } from '~/lib/queries/keys';
 import { weekIdToMonday, getTodayDayOfWeek } from '~/lib/utils/date';
 import { isCompetitionWeek } from '~/lib/utils/goals';
-import { computeWeekStats, augmentSessionsWithGarmin } from '~/lib/utils/week-view';
+import { computeWeekStats } from '~/lib/utils/week-view';
+import { augmentSessionsWithGarmin } from '~/lib/utils/junction-poc';
 import { computeSportBreakdown } from '~/lib/utils/analytics';
 import { StravaActionsBar } from '~/components/calendar/StravaActionsBar';
 import { useWeekView } from '~/lib/hooks/useWeekView';
